Convert health from int to double (1.6 changes)

+++ src/info/plugmania/mazemania/helpers/PlayerStore.java
@@ -35,6 +35,6 @@ public ItemStack[] openChest;
 
 	public GameMode gm;
 
-	public int health;
+	public double health;
 	public int hunger;
 }
